_generator/pyproject.toml
[tool.poetry]
name = "generator"
version = "0.1.0"
description = "A JD index generator"
authors = ["Sponge Team"]
license = "MIT"
readme = "../README.md"
packages = [{include = "jdgenerator"}]

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = ">=3.0.0,<4.0.0"
semver = ">=2.13.0,<3.0.0"
"ruamel.yaml" = ">=0.17"

[tool.poetry.scripts]
"generate-jd" = 'jdgenerator.generate:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
